+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-## 2023-03-22
+## 2023-03-15
+
+- info on how to wrap an R function and provide a "library" of R functions
+
+## 2023-03-12
 
 ### Added
 
@@ -9,6 +13,6 @@
 - how to see what packages are available
 - information on the MIME type necessary to serve JavaScript modules properly.
 
-## 2023-03-21
+## 2023-03-11
 
 Initial version.